Semester 4/Task 9/requirements.txt
Eel
playsound==1.2.2
pyttsx3
SpeechRecognition
PyAudio
python-dotenv
mysql-connector-python
openai
pvporcupine==1.9.5
pip
transformers
torch
accelerate
huggingface-hub
sentencepiece
pywhatkit
flask

